+++ ios/theolive/THEOplayerRCTSourceDescriptionBuilder+Theolive.swift
@@ -14,10 +14,10 @@ extension THEOplayerRCTSourceDescriptionBuilder {
      Builds a THEOplayer SourceDescription that can be passed as a source for the THEOplayer.
      - returns: a THEOlive TypedSource.
      */
-    static func buildTHEOliveDescription(_ theoliveData: [String:Any]) -> TypedSource? {
+    static func buildTHEOliveDescription(_ theoliveData: [String:Any], contentPotection: MultiplatformDRMConfiguration?) -> TypedSource? {
 #if canImport(THEOplayerTHEOliveIntegration)
       if let src = theoliveData[SD_PROP_SRC] as? String {
-          return TheoLiveSource(channelId: src)
+          return TheoLiveSource(channelId: src, drm: contentPotection)
       }
 #endif
         return nil